refactor: improve stats state management
Stats were being used inconsistently. Now document_stats holds information about the document stats and workspace_stats the same but for the entire workspace. workspace_stats.total is still incorrect because of the hardcoded 10 for text entries per document.

The bug about the negative labelling not updating the negative labelling count for the workspace is solved by this refactor, because this component gets re-rendered when the workspace redux state changes. What was causing the issue was the workspace.neg_label_num not present in the dependecy list of the removed React.useEffect().