fix: When selecting a dataset in workspace config it shows 0 then fails
This commit includes the following:
1. Removing the causing problem (trying to clear the fields ) for this bug
2. Adding a clearing functionality for dataset textfield, combo and the file name.
3. Customizing the toast notification:
a. show only one toast notification at a time and remove it from the queue, to prevent the possibility for the user to click on different buttons multiple times and cause errors.
b. customizing the error messages to reflect the error messages the frontend is receiving from the backend.
c. disabling submit buttons during the toast notification display
4. Generalizing the button IBM component for reusability